fix: add AI Observability checks (#12668)
